# apps/exposed-api/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Record_any.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ResponsePayload": {
				"properties": {
					"code": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/Record_string.any_"
					},
					"message": {
						"type": "string"
					},
					"pagination": {
						"$ref": "#/components/schemas/Record_any.any_"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse": {
				"properties": {
					"resCode": {
						"type": "number",
						"format": "double"
					},
					"resData": {
						"$ref": "#/components/schemas/ResponsePayload"
					}
				},
				"required": [
					"resCode",
					"resData"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponseStatus.Success": {
				"enum": [
					"success"
				],
				"type": "string"
			},
			"JsonObject": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"$ref": "#/components/schemas/JsonValue"
				}
			},
			"JsonValue": {
				"anyOf": [
					{
						"type": "string"
					},
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "boolean"
					},
					{
						"$ref": "#/components/schemas/JsonObject"
					},
					{
						"$ref": "#/components/schemas/JsonArray"
					}
				],
				"nullable": true
			},
			"JsonArray": {
				"items": {
					"$ref": "#/components/schemas/JsonValue"
				},
				"type": "array"
			},
			"SuccessResponseMeta": {
				"properties": {
					"nextCursor": {
						"type": "string",
						"nullable": true
					}
				},
				"additionalProperties": {
					"$ref": "#/components/schemas/JsonValue"
				},
				"required": [
					"nextCursor"
				],
				"type": "object"
			},
			"SuccessResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ApiResponseStatus.Success"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/JsonObject"
							}
						],
						"nullable": true
					},
					"meta": {
						"$ref": "#/components/schemas/SuccessResponseMeta"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponseStatus.Fail": {
				"enum": [
					"fail"
				],
				"type": "string"
			},
			"FailResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ApiResponseStatus.Fail"
					},
					"data": {
						"$ref": "#/components/schemas/JsonObject"
					}
				},
				"required": [
					"status",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponseStatus.Error": {
				"enum": [
					"error"
				],
				"type": "string"
			},
			"ErrorResponse": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/ApiResponseStatus.Error"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/JsonObject"
					}
				},
				"required": [
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CallerType": {
				"enum": [
					"anonymous",
					"user"
				],
				"type": "string"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "exposed-api",
		"version": "1.1.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/v1/games/{leagueId}": {
			"get": {
				"operationId": "GetGamesByLeagueId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Games"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "leagueId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "apppubkey",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorizationtype",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/challenges/categories/games/{gameId}": {
			"get": {
				"operationId": "GetCategoriesInGame",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ApiResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Challenges"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "gameId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v2/games": {
			"get": {
				"operationId": "GetGames",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse"
										},
										{
											"$ref": "#/components/schemas/FailResponse"
										},
										{
											"$ref": "#/components/schemas/ErrorResponse"
										}
									]
								},
								"examples": {
									"Success with Data": {
										"value": {
											"status": "success",
											"data": {
												"games": []
											},
											"meta": {
												"nextCursor": "dXNlcjpXMDdRQ1JQQTQ="
											}
										}
									},
									"Success with No Data": {
										"value": {
											"status": "success",
											"data": null
										}
									}
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"4XX": {
						"description": "Fetched Games Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FailResponse"
								}
							}
						}
					}
				},
				"description": "Retrieves a list of games based on filters, sorting, and field selection.",
				"tags": [
					"Games",
					"v2"
				],
				"security": [],
				"parameters": [
					{
						"description": "A UUID v4 for tracking the request across services.",
						"in": "header",
						"name": "Correlation-Id",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "813322a1-fd62-40f2-93a2-b3992c048012"
					},
					{
						"description": "Specifies the caller of the API.",
						"in": "header",
						"name": "Caller",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/CallerType"
						}
					},
					{
						"description": "A unique identifier for the caller. For anonymous users, provide a randomly generated UUID; for authenticated users, provide their user ID.",
						"in": "header",
						"name": "Caller-Id",
						"required": true,
						"schema": {
							"type": "string"
						},
						"example": "6c797801-2d21-4f34-ab27-e2085bf7b24d"
					},
					{
						"description": "A filter string used to narrow down the results.\n### Syntax:\n`filter=field:operator:value`\n### Supported Fields:\n- `sportId:<operator>:<number>`: Filters results by sport ID.\n- `leagueId:<operator>:<number>`: Filters results by league ID.\n- `search:<operator>:<string>`: Search by team names or the team abbreviation.\n### Supported Operators:\n- `eq`: Equals\n- `gt`: Greater than\n- `gte`: Greater than or equal\n- `has`: Contains the specified string\n- `lt`: Less than\n- `lte`: Less than or equal\n- `in`: Matches any value from a list\n### Notes:\n- Multiple filters can be combined by separating them with a comma. For example:\n`filter=leagueId:eq:1,sportId:in:[1,2,3]`\n- Ensure values are correctly encoded when used in URLs.",
						"in": "query",
						"name": "filter",
						"required": false,
						"schema": {
							"type": "string"
						},
						"examples": {
							"Example 1": {
								"value": "filter=leagueId:eq:1"
							},
							"Example 2": {
								"value": "filter=leagueId:eq:1,sportId:in:[1,2,3]"
							}
						}
					},
					{
						"description": "Defines the sorting of results.\n### Syntax:\n`sort=field:direction`\n### Supported Fields:\n- `startTsUtc`: Sort games by their start timestamp (Unix timestamp in UTC).\n### Directions:\n- `asc`: Sort in ascending order.\n- `desc`: Sort in descending order.",
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "sort=startTsUtc:desc"
					},
					{
						"description": "Define the maximum number of items to return. Defaults to 25 if not specified.",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						},
						"example": "limit=50"
					},
					{
						"description": "Opaque string value for pagination to retrieve additional results.\n### Notes:\n- Cursor strings typically end with the = character. When presenting this value as a URL or POST\nparameter, it must be encoded as `%3D`.",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						},
						"example": "cursor=eyJvZmZzZXQiOjI1LCJpZCI6MzQ1fQ%3D"
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000",
			"description": "Local development server"
		}
	]
}